+++ src/entities/course/index.ts
@@ -1,0 +1,2 @@
+export * from './api'
+export * from './model'
