+++ src/entities/course/model/type.ts
