+++ src/entities/user/api/endpoint.ts
@@ -1,10 +1,14 @@
 export const USER_API = {
   base: '/users',
-  baseRegions: '/regions/preferences',
+  baseCourse: '/courses/users',
   me: () => `${USER_API.base}/me`,
   user: (id: string) => `${USER_API.base}/${id}`,
   profile: () => `${USER_API.base}/profile`,
-  likeRegions: () => `${USER_API.baseRegions}/preferences`,
+
+  likeRegions: () => '/regions/preferences',
   deleteLikeRegions: (id: string) =>
-    `${USER_API.baseRegions}/preferences/${id}`,
+    `${USER_API.likeRegions}/preferences/${id}`,
+
+  courses: (id: string) => `${USER_API.baseCourse}/${id}`,
+  likeCourses: (id: string) => `${USER_API.baseCourse}/${id}/like`,
 }
