+++ src/entities/user/api/index.ts
@@ -7,3 +7,5 @@ export * from './update-user-profile'
 export * from './get-my-regions-preferences'
 export * from './post-my-regions-preferences'
 export * from './delete-my-region-preferences'
+export * from './get-my-like-courses'
+export * from './get-user-courses'
