+++ src/features/main/section-like-course.tsx
@@ -3,10 +3,9 @@
 import Link from 'next/link'
 import Spacer from '@/src/shared/ui/Spacer'
 import GridCourse from '@/src/features/course/grid-course'
-import { CourseType } from '@/src/entities/course/type'
+import { CourseType } from '@/src/entities/course'
 import NoLikedCourse from '@/src/shared/ui/NoLikedCourse'
-import { useGetMyProfile } from '@/src/entities/user'
-import { useGetLikeCourses } from '@/src/entities/course/query'
+import { useGetMyProfile, useGetLikeCourses } from '@/src/entities/user'
 import { getLoginUrl } from '@/src/entities/login/api'
 import { useRouter } from 'next/navigation'
 import { ChevronRight } from 'lucide-react'
