+++ src/views/detail-course/index.tsx
@@ -5,7 +5,7 @@ import ProfileImage from '@/src/shared/ui/ProfileImage'
 import Spacer from '@/src/shared/ui/Spacer'
 import CoursePlanDetailLayout from '@/src/widgets/course-plan-detail-layout'
 import { passFromCreate } from '@/src/shared/utils/date'
-import { useGetCourse } from '@/src/entities/course/query'
+import { useGetCourse } from '@/src/entities/course'
 import { useGetComments } from '@/src/entities/comment/query'
 import defaultImg from '@/src/assets/images/(logo)/logo_default.png'
 import ReviewCommentCard from '@/src/widgets/review-comment-card'
