+++ src/views/main-course/index.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import { CourseType } from '@/src/entities/course/type'
+import { CourseType } from '@/src/entities/course'
 import CardCourseList from '@/src/features/course/card-list-course'
 import Spacer from '@/src/shared/ui/Spacer'
 import FloatingWriteButton from '@/src/widgets/floating-write-btn'
 import { useRouter } from 'next/navigation'
 import { Fragment, useEffect } from 'react'
-import { useGetCourses } from '@/src/entities/course/query'
+import { useGetCourses } from '@/src/entities/course'
 import { useGetMyLikeRegions } from '@/src/entities/user'
 import RegionCascaderWithLikes from '@/src/shared/ui/RegionCascader'
 import useRegionStore from '@/src/shared/store/regionStore'
