+++ src/widgets/course-plan-form-layout/index.tsx
@@ -7,16 +7,16 @@ import Spacer from '@/src/shared/ui/Spacer'
 import SearchPlace from '@/src/views/search-place'
 import Header from '@/src/widgets/header'
 import type { CoursePlanPlaceType } from '@/src/entities/place/type'
-import type { CoursePayloadType } from '@/src/entities/course/type'
-import { COURSE_QUERY_KEY } from '@/src/entities/course/query'
+import type { CoursePayloadType } from '@/src/entities/course'
 import { PLAN_QUERY_KEY } from '@/src/entities/plan/query'
 import FormSections from '@/src/features/course/form-course'
 import { useForm } from 'react-hook-form'
 import {
-  useCreateCourse,
+  COURSE_QUERY_KEY,
   useGetCourse,
+  usePostCourse,
   useUpdateCourse,
-} from '@/src/entities/course/query'
+} from '@/src/entities/course'
 import {
   useCreatePlan,
   useGetPlan,
@@ -154,7 +154,7 @@ export default function CoursePlanFormLayout({
     }
   }, [level, type, isDataLoaded])
 
-  const { mutate: courseMutate } = useCreateCourse()
+  const { mutate: courseMutate } = usePostCourse()
   const { mutate: planMutate } = useCreatePlan()
   const { mutate: courseUpdateMutate } = useUpdateCourse(id || '')
   const { mutate: planUpdateMutate } = useUpdatePlan(id || '')
